requirements.txt
fastapi
uvicorn
requests
python-dotenv
supabase
httpx>=0.23.0